requirements.txt
dotenv
langchain-core
fastapi
uvicorn
pydantic
streamlit
requests
pillow
pdfplumber
python-docx
beautifulsoup4
unstructured
chromadb
qdrant-client
tavily-python
google-generativeai
openai
nest_asyncio
langchain
langchain-openai
langchain-community
langchain-google-genai
langchain-tavily
langchain-qdrant
langgraph
langsmith
python-multipart
langchain_perplexity
sounddevice
silero_vad
pyaudio
numpy
torch
websockets
python-socketio